WIP debugging

+++ src/sdp_solve/SDP_Solver/run/step/initialize_schur_complement_solver/bigint_syrk/BigInt_Shared_Memory_Syrk_Context.cxx
@@ -58,7 +58,7 @@ BigInt_Shared_Memory_Syrk_Context::BigInt_Shared_Memory_Syrk_Context(
 void BigInt_Shared_Memory_Syrk_Context::bigint_syrk_blas(
   El::UpperOrLower uplo,
   const std::vector<El::DistMatrix<El::BigFloat>> &bigint_input_matrix_blocks,
-  const std::vector<size_t> &block_indices_per_shared_memory_comm,
+  const std::vector<int> &block_indices_per_shared_memory_comm,
   El::DistMatrix<El::BigFloat> &bigint_output, Timers &timers)
 {
   Scoped_Timer timer(timers, "bigint_syrk_blas");
